week3 - [refactor] move all input/output elements from LottoMain.kt to LottoUI.kt
